Walking. It's there, and it works.
HAHAHA. Bug notice: due to how I work the camera, canvas clearing on the
top left edge of the map. Also, the steps are kinda twitchy. I'm not
sure what exactly is causing this, it could be anywhere from timing to
getting multiple values ( increasing and decreasing ) position values or
perhaps doing things multiple times per frame or even that stuff is done
out of order. Either way, you start stop and change directions similar
to how you would in a legit game so that is good.