Setting up for a message-log.
I plan to utilize this in scripting. Also I refactored a bit of code for
the sake of consistent & decent coding practices and since the computer
I'm developing on has a crappy built in graphics processor, and I just
removed my graphics card I went ahead and removed some css effects. I
plan to put them back in but for now I changed them to maintain my
patience with this.